Split binaries and library from each other
Makes it whole lot simpler no need to check for K8S env variable
