+++ Cargo.toml
@@ -23,6 +23,14 @@ either = "1.6.1"
 validator = { version = "0.14.0", features = ["derive"] }
 thiserror = "1.0.29"
 
-[[bin]]
+[lib]
 name = "cloudflare"
-path = "src/cloudflare.rs"
+path = "src/lib.rs"
+
+[[bin]]
+name = "client"
+path = "src/client/main.rs"
+
+[[bin]]
+name = "controller"
+path = "src/controller/main.rs"
