+++ examples/k8s/dns.yaml
@@ -44,10 +44,3 @@ spec:
       secretKeyRef:
         name: cloudflare
         key: token
----
-apiVersion: v1
-kind: Secret
-metadata:
-  name: cloudflare
-data:
-  token: <BASE64 ENCODED STRING>
